landscape: show all invites in inbox
Fixes #3925.
